fpga: Add flag to indicate bitstream needs decrypting
Add a flag that is passed to the write_init() callback, indicating
that the bitstream is encrypted.

The low-level driver will deal with the flag, or return an error,
if encrypted bitstreams are not supported.

Signed-off-by: Moritz Fischer <mdf@kernel.org>
Acked-by: Michal Simek <michal.simek@xilinx.com>
Signed-off-by: Alan Tull <atull@kernel.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 0f4f0c8ff1da9171bca0dc01ce5551e8b6d2f0f3)

Change-Id: I048d4ac5cd8be763a58cae36ddf94da18779cd39