fpga: machxo2: Allow infinite waits until not busy
The erase sequence can take much longer than most commands. As the
command can take variable amounts of time for different parts, having a
single maximum number of times to check for not busy is brittle.

This commit allows passing 0 for the number of maximum loops to denote
no maximum. This should be fine as at this point in the programming
sequence we know we are talking to a valid chip, and if an error is hit,
the chipo will disable the busy flag and set an error condition.

Change-Id: I6dab3659314f98fa7b8aa0bb251067d3fec58ba9
Signed-off-by: Gabriel Smith <gsmith@d3engineering.com>