fpga: of-region: Support statically loading firmware
This adds support for specifying a firmware image in a static DTB rather
than relying on overlays. Due to how the of-fpga mechanism works, the
fpga will not be able to be reprogrammed at runtime with a different
image as there is no overlay to remove to remove the existing firmware.

Change-Id: I7803129b8c7cd01c83abc0355e6318de1852ebbe
Signed-off-by: Gabriel Smith <gsmith@d3engineering.com>