timestamp: use R32.2.1 version of timekeeping.c to effectively remove patch
In earlier D3 releases we applied a patch provided by Nvidia that
published an offset. That code has since been applied in Nvidia's
official release but it is not in the same form as the patch that we
had applied. This commit brings us in-sync with Nvidia. We are not
losing functionality and we are closer to their official code.